fix(mirror): Added mirroring feature
Added feature to submit full sync data packet to secondary server when
MESH_MIRROR env variable set.

Feature is VERY alpha and has no added security or limitation on data
sent to the mirror server, in effect its the same as if the client had
directly posts to the mirror - this feature is best used with a live and
development server.